+++ Generator~/WallstopStudios.UnityHelpers.Proto.Generator/AssemblyInfo.cs
@@ -1,0 +1,6 @@
+// MIT License - Copyright (c) 2026 wallstop
+// Full license text: https://github.com/wallstop/unity-helpers/blob/main/LICENSE
+
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("WallstopStudios.UnityHelpers.Proto.Generator.Tests")]